768 - Change price columns to BIGINT / INT8RANGE types

See merge request vng/diwi!740

+++ backend/src/main/java/nl/vng/diwi/dal/entities/HouseblockSnapshotSqlModel.java
@@ -115,14 +115,14 @@ public class HouseblockSnapshotSqlModel {
         private UUID ownershipId;
         private OwnershipType ownershipType;
         private Integer ownershipAmount;
-        private Integer ownershipValue;
-        private Integer ownershipRentalValue;
+        private Long ownershipValue;
+        private Long ownershipRentalValue;
         private UUID ownershipRangeCategoryId;
         private UUID ownershipRentalRangeCategoryId;
-        private Integer ownershipValueRangeMin;
-        private Integer ownershipValueRangeMax;
-        private Integer ownershipRentalValueRangeMin;
-        private Integer ownershipRentalValueRangeMax;
+        private Long ownershipValueRangeMin;
+        private Long ownershipValueRangeMax;
+        private Long ownershipRentalValueRangeMin;
+        private Long ownershipRentalValueRangeMax;
     }
 
 }
